www.norman-databases.org

+++ database/migrations/2025_08_26_085300_add_substance_id_to_ecotox_pnec3_table.php
@@ -13,7 +13,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('ecotox_pnec3', function (Blueprint $table) {
-            $table->foreignId('substance_id')->nullable()->default(null)->references('id')->on('susdat_substances')->onUpdate('cascade')->onDelete('restrict');
+            if (!Schema::hasColumn('ecotox_pnec3', 'substance_id')) {
+                $table->foreignId('substance_id')->nullable()->default(null)->references('id')->on('susdat_substances')->onUpdate('cascade')->onDelete('restrict');
+            }
         });
     }
